requirements.txt
pytelegrambotapi==4.11.0
pymupdf==1.23.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==9.5.0
requests==2.31.0
